refactor: Change file names according to pep 8 standards
