+++ exam_django/asset/service/asset_approve_service/asset_approve_service.py
@@ -2,7 +2,7 @@ from rest_framework.renderers import JSONRenderer
 from rest_framework import status
 
 from asset.models import Asset
-from asset.serializers.AssetSerializer import AssetReadSerializer
+from asset.serializers.asset_serializer import AssetReadSerializer
 from notification.service.EmailService import EmailService
