+++ exam_django/asset/views/location_view.py
